# refactor: faster filtering algorithm for null consts