--- a/benches/elgamal.rs
+++ b/benches/elgamal.rs
@@ -5,8 +5,8 @@ use ezkl::circuit::modules::elgamal::{
 };
 use ezkl::circuit::modules::Module;
 use ezkl::circuit::*;
-use ezkl::execute::create_proof_circuit_kzg;
 use ezkl::pfsys::create_keys;
+use ezkl::pfsys::create_proof_circuit_kzg;
 use ezkl::pfsys::srs::gen_srs;
 use ezkl::pfsys::TranscriptType;
 use ezkl::tensor::*;