--- a/src/circuit/ops/layouts.rs
+++ b/src/circuit/ops/layouts.rs
@@ -7,7 +7,8 @@ use std::{
 use halo2_proofs::circuit::Value;
 use halo2curves::ff::PrimeField;
 use itertools::Itertools;
-use log::error;
+use log::{debug, error, trace};
+use rayon::slice::ParallelSliceMut;
 
 use super::{
     chip::{BaseConfig, CheckMode, CircuitError},
@@ -57,14 +58,24 @@ pub fn dot<F: PrimeField + TensorType + PartialOrd>(
     region: &mut RegionCtx<F>,
     values: &[ValTensor<F>; 2],
 ) -> Result<ValTensor<F>, Box<dyn Error>> {
+    // time this entire function run
+    let global_start = std::time::Instant::now();
+
     let mut values = values.clone();
 
-    let mut removal_indices: HashSet<usize> = HashSet::new();
-    removal_indices.extend(values[0].get_const_zero_indices()?);
-    removal_indices.extend(values[1].get_const_zero_indices()?);
-    let removal_indices = removal_indices.into_iter().collect_vec();
-    values[0].remove_indices(&removal_indices)?;
-    values[1].remove_indices(&removal_indices)?;
+    // this section has been optimized to death, don't mess with it
+    let mut removal_indices = values[0].get_const_zero_indices()?;
+    let second_zero_indices = values[1].get_const_zero_indices()?;
+    removal_indices.extend(second_zero_indices);
+    removal_indices.par_sort_unstable();
+    removal_indices.dedup();
+
+    // is already sorted
+    values[0].remove_indices(&mut removal_indices, true)?;
+    values[1].remove_indices(&mut removal_indices, true)?;
+
+    let elapsed = global_start.elapsed();
+    trace!("filtering const zero indices took: {:?}", elapsed);
 
     if values[0].len() != values[1].len() {
         return Err(Box::new(TensorError::DimMismatch("dot".to_string())));
@@ -81,9 +92,12 @@ pub fn dot<F: PrimeField + TensorType + PartialOrd>(
         region.increment(overflowed_len);
         let num_constants = values[0].num_constants() + values[1].num_constants();
         region.increment_constants(num_constants);
+        trace!("dummy dot layout took {:?}", global_start.elapsed());
+
         return Ok(Tensor::from([ValType::<F>::from(Value::<F>::unknown())].into_iter()).into());
     }
 
+    let start = std::time::Instant::now();
     let mut inputs = vec![];
     let mut assigned_len = 0;
     for (i, input) in values.iter().enumerate() {
@@ -95,16 +109,25 @@ pub fn dot<F: PrimeField + TensorType + PartialOrd>(
         };
         inputs.push(inp);
     }
+    let elapsed = start.elapsed();
+    trace!("assigning inputs took: {:?}", elapsed);
 
     // Now we can assign the dot product
+    // time this step
+    let start = std::time::Instant::now();
     let accumulated_dot = accumulated::dot(&[inputs[0].clone(), inputs[1].clone()])
         .expect("accum poly: dot op failed");
+    let elapsed = start.elapsed();
+    trace!("calculating accumulated dot took: {:?}", elapsed);
 
+    let start = std::time::Instant::now();
     let (output, output_assigned_len) = region.assign_with_duplication(
         &config.output,
         &accumulated_dot.into(),
         &config.check_mode,
     )?;
+    let elapsed = start.elapsed();
+    trace!("assigning output took: {:?}", elapsed);
 
     assert_eq!(assigned_len, output_assigned_len);
 
@@ -141,6 +164,10 @@ pub fn dot<F: PrimeField + TensorType + PartialOrd>(
 
     region.increment(assigned_len);
     // last element is the result
+
+    let elapsed = global_start.elapsed();
+    trace!("dot layout took: {:?}", elapsed);
+    trace!("----------------------------");
     Ok(last_elem)
 }
 
@@ -599,6 +626,9 @@ pub fn pairwise<F: PrimeField + TensorType + PartialOrd>(
     values: &[ValTensor<F>; 2],
     op: BaseOp,
 ) -> Result<ValTensor<F>, Box<dyn Error>> {
+    // time to calculate the value of the output
+    let global_start = std::time::Instant::now();
+
     let (mut lhs, mut rhs) = (values[0].clone(), values[1].clone());
 
     let broadcasted_shape = get_broadcasted_shape(lhs.dims(), rhs.dims())?;
@@ -608,19 +638,16 @@ pub fn pairwise<F: PrimeField + TensorType + PartialOrd>(
     let orig_lhs = lhs.clone();
     let orig_rhs = rhs.clone();
 
-    let zero_indices_a = lhs.get_const_zero_indices()?;
-    let zero_indices_b = rhs.get_const_zero_indices()?;
-    let zero_indices = zero_indices_a
-        .iter()
-        .chain(zero_indices_b.iter())
-        .cloned()
-        .collect::<HashSet<_>>()
-        .iter()
-        .cloned()
-        .collect::<Vec<_>>();
+    let first_zero_indices = lhs.get_const_zero_indices()?;
+    let second_zero_indices = rhs.get_const_zero_indices()?;
+    let mut removal_indices = first_zero_indices.clone();
+    removal_indices.extend(second_zero_indices.clone());
+    removal_indices.par_sort_unstable();
+    removal_indices.dedup();
 
-    lhs.remove_indices(&zero_indices)?;
-    rhs.remove_indices(&zero_indices)?;
+    // is already sorted
+    lhs.remove_indices(&mut removal_indices, true)?;
+    rhs.remove_indices(&mut removal_indices, true)?;
 
     if lhs.len() != rhs.len() {
         return Err(Box::new(CircuitError::DimMismatch(format!(
@@ -636,20 +663,29 @@ pub fn pairwise<F: PrimeField + TensorType + PartialOrd>(
         let vals = vec![ValType::Value(Value::<F>::unknown()); broadcasted_shape.iter().product()];
         let mut tensor: Tensor<ValType<F>> = Tensor::from(vals.into_iter());
         tensor.reshape(&broadcasted_shape);
+
+        trace!(
+            "dummy pairwise {} layout took {:?}",
+            op.as_str(),
+            global_start.elapsed()
+        );
+
         return Ok(tensor.into());
     }
 
     let mut inputs = vec![];
-
     for (i, input) in [lhs.clone(), rhs.clone()].iter().enumerate() {
         let inp = {
             let res = region.assign(&config.inputs[i], input)?;
+
             res.get_inner()?
         };
         inputs.push(inp);
     }
 
     // Now we can assign the dot product
+    // time the calc
+    let start = std::time::Instant::now();
     let op_result = match op {
         BaseOp::Add => add(&inputs),
         BaseOp::Sub => sub(&inputs),
@@ -660,6 +696,8 @@ pub fn pairwise<F: PrimeField + TensorType + PartialOrd>(
         error!("{}", e);
         halo2_proofs::plonk::Error::Synthesis
     })?;
+    let elapsed = start.elapsed();
+    debug!("pairwise {} calc took {:?}", op.as_str(), elapsed);
 
     let output = region.assign(&config.output, &op_result.into())?;
 
@@ -677,11 +715,11 @@ pub fn pairwise<F: PrimeField + TensorType + PartialOrd>(
     // infill the zero indices with the correct values from values[0] or values[1]
     let mut actual_output: ValTensor<F> =
         Into::<Tensor<ValType<F>>>::into((0..broadcasted_shape.iter().product()).map(|i| {
-            if zero_indices.contains(&i) {
+            if removal_indices.contains(&i) {
                 let a = orig_lhs.get_inner_tensor().unwrap()[i].clone();
                 let b = orig_rhs.get_inner_tensor().unwrap()[i].clone();
-                let a_is_null = zero_indices_a.contains(&i);
-                let b_is_null = zero_indices_b.contains(&i);
+                let a_is_null = first_zero_indices.contains(&i);
+                let b_is_null = second_zero_indices.contains(&i);
                 let both_null = a_is_null && b_is_null;
 
                 match op {
@@ -724,6 +762,9 @@ pub fn pairwise<F: PrimeField + TensorType + PartialOrd>(
         .into();
 
     actual_output.reshape(&broadcasted_shape)?;
+
+    let end = global_start.elapsed();
+    trace!("pairwise {} layout took {:?}", op.as_str(), end);
 
     Ok(actual_output)
 }
@@ -774,6 +815,7 @@ pub fn neg<F: PrimeField + TensorType + PartialOrd>(
 ) -> Result<ValTensor<F>, Box<dyn Error>> {
     let input = {
         let res = region.assign(&config.inputs[1], &values[0])?;
+
         res.get_inner()?
     };
 
@@ -1101,7 +1143,7 @@ pub fn conv<F: PrimeField + TensorType + PartialOrd + std::marker::Send + std::m
     }
 
     // increment the region
-    region.increment(std::cmp::max(assigned_image_len, assigned_kernel_len));
+    region.increment(std::cmp::max(assigned_kernel_len, assigned_image_len));
 
     let og_dims = image.dims().to_vec();
 
@@ -1470,7 +1512,7 @@ pub fn downsample<F: PrimeField + TensorType + PartialOrd>(
     let processed_output =
         tensor::ops::downsample(&input.get_inner_tensor()?, *axis, *stride, *modulo)?;
     let output = region.assign(&config.output, &processed_output.into())?;
-    region.increment(input.len());
+    region.increment(std::cmp::max(input.len(), output.len()));
     Ok(output)
 }
 
@@ -1504,6 +1546,9 @@ pub fn nonlinearity<F: PrimeField + TensorType + PartialOrd>(
     values: &[ValTensor<F>; 1],
     nl: &LookupOp,
 ) -> Result<ValTensor<F>, Box<dyn Error>> {
+    // time the entire operation
+    let timer = std::time::Instant::now();
+
     let x = &values[0];
 
     if region.is_dummy() {
@@ -1513,6 +1558,13 @@ pub fn nonlinearity<F: PrimeField + TensorType + PartialOrd>(
         let vals = vec![ValType::Value(Value::<F>::unknown()); x.len()];
         let mut x = Tensor::from(vals.into_iter());
         x.reshape(dims);
+        // total time taken
+        trace!(
+            "dummy nonlinearity {} layout took {:?}",
+            <LookupOp as Op<F>>::as_string(nl),
+            timer.elapsed()
+        );
+
         return Ok(x.into());
     }
 
@@ -1534,7 +1586,9 @@ pub fn nonlinearity<F: PrimeField + TensorType + PartialOrd>(
 
     let mut output = region.assign(&config.lookup_output, &output.into())?;
 
-    (0..x.len()).for_each(|i| {
+    assert_eq!(w.len(), output.len());
+
+    (0..output.len()).for_each(|i| {
         let (x, y) = config.lookup_input.cartesian_coord(region.offset() + i);
         let selector = config.lookup_selectors.get(&(nl.clone(), x));
         region.enable(selector, y).unwrap();
@@ -1542,7 +1596,14 @@ pub fn nonlinearity<F: PrimeField + TensorType + PartialOrd>(
 
     output.reshape(x.dims())?;
 
-    region.increment(x.len());
+    region.increment(output.len());
+
+    let elapsed = timer.elapsed();
+    trace!(
+        "nonlinearity {} layout took {:?}",
+        <LookupOp as Op<F>>::as_string(nl),
+        elapsed
+    );
 
     // constrain the calculated output to a column
     Ok(output)
@@ -1610,8 +1671,8 @@ pub fn max<F: PrimeField + TensorType + PartialOrd>(
         Some(i) => Tensor::new(Some(&[Value::known(i128_to_felt::<F>(i))]), &[1])?.into(),
     };
 
-    let assigned_max_val = region.assign(&config.inputs[1], &max_val)?;
-    region.next();
+    let assigned_max_val: ValTensor<F> = region.assign(&config.inputs[1], &max_val)?;
+    region.increment(assigned_max_val.len());
 
     let unit: ValTensor<F> =
         Tensor::from(vec![region.assign_constant(&config.inputs[1], F::from(1))?].into_iter())
@@ -1705,7 +1766,7 @@ pub fn min<F: PrimeField + TensorType + PartialOrd>(
     };
 
     let assigned_min_val = region.assign(&config.inputs[1], &min_val)?;
-    region.next();
+    region.increment(assigned_min_val.len());
 
     let unit: ValTensor<F> =
         Tensor::from(vec![region.assign_constant(&config.inputs[1], F::from(1))?].into_iter())