Improves layout speed by a factor of 20 for dummy layouts and by a factor of 10 for "real" layouts. 

Prior setup phase: 

```bash 
[T] [1s, ezkl::circuit::ops::layouts] - filtering const zero indices took: 1.154125ms
[T] [1s, ezkl::circuit::ops::layouts] - assigning inputs took: 53.292µs
[T] [1s, ezkl::circuit::ops::layouts] - calculating accumulated dot took: 2.208µs
[T] [1s, ezkl::circuit::ops::layouts] - assigning output took: 23.167µs
[T] [1s, ezkl::circuit::ops::layouts] - dot layout took: 1.280833ms
```

New setup phase:  

```bash
[T] [1s, ezkl::circuit::ops::layouts] - filtering const zero indices took: 18.708µs
[T] [1s, ezkl::circuit::ops::layouts] - assigning inputs took: 66.917µs
[T] [1s, ezkl::circuit::ops::layouts] - calculating accumulated dot took: 3.833µs
[T] [1s, ezkl::circuit::ops::layouts] - assigning output took: 20.5µs
[T] [1s, ezkl::circuit::ops::layouts] - dot layout took: 154.041µs
```